+++ src/app/dashboard/general-module/general-module.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { FormGroup, FormControl, Validators, FormArray } from '@angular/forms';
 import { AuthService } from '../../services/auth.service';
 import { ActivatedRoute } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
@@ -13,6 +13,8 @@ import { GuildService } from '../../services/guild.service';
 })
 export class GeneralModuleComponent extends ModuleConfig implements OnInit {
   moduleName = 'general';
+
+  reactionRolesIndices = [0,1,2,3];
 
   get general() { return this.savedGuild.general; }
 
@@ -28,14 +30,27 @@ export class GeneralModuleComponent extends ModuleConfig implements OnInit {
   }
 
   buildForm() {
-    return new FormGroup({
+    const fg = new FormGroup({
       prefix: new FormControl('', [
         Validators.required, 
         Validators.maxLength(5) 
       ]),
-      ignoredChannels: new FormControl(),
-      autoRoles: new FormControl()
+      ignoredChannels: new FormControl([]),
+      autoRoles: new FormControl([]),
+      reactionRoles: new FormArray([])
     });
+
+    for (const i of this.reactionRolesIndices) {
+      (fg.get('reactionRoles') as FormArray).setControl(i,
+        new FormGroup({
+          channel: new FormControl(''),
+          role: new FormControl(''),
+          emote: new FormControl('', 
+            Validators.pattern(/(\u00a9|\u00ae|[\u2000-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff]){1}/gm)),
+          messageId: new FormControl('', Validators.pattern(/[0-9]{18}/g))
+        }));    
+    }
+    return fg;
   }
   
   initFormValues(savedGuild: any) {
@@ -43,5 +58,10 @@ export class GeneralModuleComponent extends ModuleConfig implements OnInit {
     this.form.controls.prefix.setValue(general.prefix);
     this.form.controls.ignoredChannels.setValue(general.ignoredChannels);
     this.form.controls.autoRoles.setValue(general.autoRoles);
+    (this.form.controls.reactionRoles as FormArray).setValue(general.reactionRoles);
+  }
+
+  addEmoji($event) {
+    console.log($event);    
   }
 }