+++ src/app/material-module.ts
@@ -24,6 +24,7 @@ import {ColorPickerModule} from 'ngx-color-picker';
 import {CdkTableModule} from '@angular/cdk/table';
 import {MatSortModule} from '@angular/material/sort';
 import {MatSliderModule} from '@angular/material/slider';
+import {PickerModule} from '@ctrl/ngx-emoji-mart';
 
 @NgModule({
   exports: [
@@ -52,6 +53,7 @@ import {MatSliderModule} from '@angular/material/slider';
     MatSlideToggleModule,
     MatSelectModule,
     MatSnackBarModule,
+    PickerModule,
     ScrollingModule
   ]
 })