+++ src/assets/docs/changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
+*Built on [2PG](https://2pg.xyz)*
+
 ---
 
-# Alpha
+## Alpha
+
+## 2.0.0a
 
 ![v2.0.0a Dashboard](assets/docs/img/dashboard-v2.0.0a.png)
 
@@ -12,3 +16,13 @@
 - migrate auto mod features
 
 **Remove**
+
+---
+
+## Older Versions
+
+### 1.x
+
+![v1.x Dashboard](assets/docs/img/dashboard-v1.x.png)
+
+![v1.x Home](assets/docs/img/dashboard-home-v1.x.png)